requirements.txt
certifi==2022.9.24
charset-normalizer==2.0.12
chart-studio==1.1.0
cmake==3.25.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
idna==3.4
imageio==2.15.0
joblib==1.1.1
kiwisolver==1.3.1
lxml==4.9.1
matplotlib==3.3.4
monai==0.8.1
networkx==2.5.1
nibabel==3.2.2
nilearn==0.9.2
numpy==1.19.5
packaging==21.3
pandas==1.1.5
patsy==0.5.3
pi==0.1.2
Pillow==8.4.0
plotly==5.11.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.1.1
PyYAML==6.0
requests==2.27.1
retrying==1.3.3
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
statsmodels==0.12.2
tenacity==8.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.1.1
urllib3==1.26.12
webcolors==1.11.1
